* Moved AtomicCounterEditor and provider to its own package * Fixed issue where atomic counters on non root nodes wouldn't work * Fixed some typos * Left some FIXMEs
